# plans/phase4/openapi_ireios4.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Real Estate AI Lead Agent",
    "description": "Advanced client-grade AI assistant backend for real estate tracking, dynamic questioning, and intelligent lead evaluation.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/events/stream": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Events Stream",
        "description": "Tenant-scoped Server-Sent Events bridge from the event bus.\n\nThe caller's ``tenant_id`` (``Client_<id>``) is used to filter envelopes;\nonly events for that tenant are streamed. Sends a ``: ping`` comment every\n``HEARTBEAT_SEC`` seconds to keep proxies from buffering/closure.",
        "operationId": "events_stream_api_v1_events_stream_get",
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/leads/{lead_id}/timeline": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Lead Timeline",
        "description": "Return envelope-shaped events for a lead, tenant-scoped.\n\nSourced from ``event_logs`` via the lead's sessions. When no logs exist\nthe list is empty (stable schema for the FE). Real producers (Phase 7/9)\nwill enrich this later.",
        "operationId": "lead_timeline_api_v1_events_leads__lead_id__timeline_get",
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/predictions/revenue": {
      "get": {
        "tags": [
          "predictions"
        ],
        "summary": "Get Revenue",
        "operationId": "get_revenue_api_v1_predictions_revenue_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/predictions/cancellation-risk": {
      "get": {
        "tags": [
          "predictions"
        ],
        "summary": "Get Cancellation Risk",
        "operationId": "get_cancellation_risk_api_v1_predictions_cancellation_risk_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/predictions/inventory": {
      "get": {
        "tags": [
          "predictions"
        ],
        "summary": "Get Inventory Prediction",
        "operationId": "get_inventory_prediction_api_v1_predictions_inventory_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/predictions/cashflow": {
      "get": {
        "tags": [
          "predictions"
        ],
        "summary": "Get Cashflow",
        "operationId": "get_cashflow_api_v1_predictions_cashflow_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      },
      "APIKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "api_key"
      }
    }
  }
}